add combined democracy migration
